Teste case, refatorando e ajustando classes.
