--- docker-compose.yml
@@ -1,11 +1,0 @@
-version: "3.3"
-
-services:
-  postgres:
-    image: postgres:14
-    ports:
-      - "5432:5432"
-    environment:
-      POSTGRES_DB: bd-test-boti
-      POSTGRES_USER: boti
-      POSTGRES_PASSWORD: banco123
