+++ src/main/kotlin/br/com/blz/testjava/dto/NewInventoryForm.kt
@@ -1,0 +1,8 @@
+package br.com.blz.testjava.dto
+
+import jakarta.validation.constraints.NotEmpty
+
+data class NewInventoryForm(
+  @field:NotEmpty(message = "Warehouses cannot be empty")
+  val warehouses: MutableList<NewWarehouseForm> = mutableListOf(),
+)
